+++ Tests/RoutingTests/AddTests.swift
@@ -1,6 +1,6 @@
 import XCTest
 import HTTP
-import HTTPRouting
+import Routing
 
 //class AddTests: XCTestCase {
 //    static var allTests = [
