+++ Tests/RoutingTests/GroupTests.swift
@@ -1,6 +1,6 @@
 import XCTest
 import HTTP
-import HTTPRouting
+import Routing
 
 //class GroupTests: XCTestCase {
 //    static var allTests = [
